:arrow_down: Change aya dependencies so everyone has the same consistent version
Instead of the git repo to pull the latest version from, I changed it to an actual version. This gives us more control over the versioning as I experienced problems on different machines. This makes also sure that we don't suddenly have a broken program when Aya creates a release with breaking changes